support switching between ogg and m4a for ios
